refactor DTOs and centralize product/category mappings

+++ src/main/java/com/michele/mocks/dto/products/CreateProductRequest.java
@@ -1,5 +1,7 @@
 package com.michele.mocks.dto.products;
 
+import java.math.BigDecimal;
+
 public record CreateProductRequest(
         String sku,
         String name,
@@ -14,7 +16,7 @@ public record CreateProductRequest(
         Double heightCm,
         Integer minQuantity,
         String imageUrl,
-        Double sellPrice,
-        Double purchasePrice,
+        BigDecimal sellPrice,
+        BigDecimal purchasePrice,
         String currency) {
 }
